requirements.txt
pygithub >= 1.58.2